+++ client/cody-shared/src/chat/recipes/vscode-recipes.ts
@@ -7,6 +7,7 @@ import { GenerateDocstring } from './generate-docstring'
 import { GenerateTest } from './generate-test'
 import { GitHistory } from './git-log'
 import { ImproveVariableNames } from './improve-variable-names'
+import { NextQuestions } from './next-questions'
 import { Recipe } from './recipe'
 import { TranslateToLanguage } from './translate'
 
@@ -36,6 +37,7 @@ function init(): void {
         new Fixup(),
         new TranslateToLanguage(),
         new FindCodeSmells(),
+        new NextQuestions(),
     ]
 
     for (const recipe of recipes) {